refactor(cactus-web): remove now-unused helpers
CACTUS-1093
