--- modules/cactus-web/src/helpers/flexItem.ts
@@ -1,5 +1,0 @@
-import { FlexboxProps, system } from 'styled-system'
-
-export type FlexItemProps = Pick<FlexboxProps, 'flex' | 'flexBasis' | 'flexGrow' | 'flexShrink'>
-
-export const flexItem = system({ flex: true, flexBasis: true, flexGrow: true, flexShrink: true })
